Back out changes from the vendor import of gc 7.2.
Estimated hours taken: 1
Branches: main

Back out changes from the vendor import of gc 7.2.

+++ boehm_gc/libatomic_ops/src/atomic_ops/sysdeps/gcc/alpha.h
@@ -27,6 +27,7 @@ AO_nop_full(void)
 {
   __asm__ __volatile__("mb" : : : "memory");
 }
+
 #define AO_HAVE_nop_full
 
 AO_INLINE void
@@ -34,6 +35,7 @@ AO_nop_write(void)
 {
   __asm__ __volatile__("wmb" : : : "memory");
 }
+
 #define AO_HAVE_nop_write
 
 /* mb should be used for AO_nop_read().  That's the default.    */
@@ -58,6 +60,7 @@ AO_compare_and_swap(volatile AO_t *addr,
                      :"=&r" (temp), "=m" (*addr), "=&r" (was_equal)
                      : "r" (new_val), "Ir" (old)
                      :"memory");
-  return (int)was_equal;
+  return was_equal;
 }
+
 #define AO_HAVE_compare_and_swap
